+++ TaskManagement-backend/routes/task.js
@@ -30,7 +30,7 @@ async function getAllTasks(req, res, next) {
   }
 
   async function getTaskById(req, res, next) {
-    logger.info('getTask with id: ', req.params.id)
+   // logger.info('getTask with id: ', req.params.id)
 
   if (!req.params.id) {
     res.status(500).send('The param id is not defined')
@@ -48,10 +48,10 @@ async function getAllTasks(req, res, next) {
   }
 
 async function createTask(req, res, next) {
-    logger.info('post task ', req.body)
+   // logger.info('post task ', req.body)
     const task = req.body
     try {
-      logger.info(' task.assigned_team ' + task.assigned_team)
+   //   logger.info(' task.assigned_team ' + task.assigned_team)
       const team = await Team.findOne({ idTeam: task.assigned_team.idTeam })
     if (!team) {
       res.status(404).send('Team not found')
@@ -62,7 +62,7 @@ async function createTask(req, res, next) {
       const userAssigned = await User.findById(task.assigned_user)
 
       if (!userAssigned) {
-        logger.info('userAssigned not found.')
+        logger.error('userAssigned not found.')
         return res.status(400).end()
       }
       req.body.assigned_user = userAssigned._id
@@ -71,16 +71,16 @@ async function createTask(req, res, next) {
     //Valido autor
     if (task.comments && Array.isArray(task.comments)) { //es una coleccion
       for (const comment of task.comments) {
-        logger.info('Validating comment author:', comment.author);
+      //  logger.info('Validating comment author:', comment.author);
         if (comment.author) {
           const userAuthor = await User.findById(comment.author);
           if (!userAuthor) {
-            logger.info('userAuthor not found for comment.');
+            logger.error('userAuthor not found for comment.');
             return res.status(400).send('Invalid comment author');
           }
           comment.author = userAuthor._id; // Reemplaza con el ID validado
         } else {
-          logger.info('Comment author is missing.');
+          logger.error('Comment author is missing.');
           return res.status(400).send('Comment author is required');
         }
       }
@@ -91,7 +91,7 @@ async function createTask(req, res, next) {
       const authorizedUser = await User.findById(task.authorized_by)
 
       if (!authorizedUser) {
-        logger.info('authorizedUser not found.')
+        logger.error('authorizedUser not found.')
         return res.status(400).end()
       }
       req.body.authorized_by = authorizedUser._id
@@ -139,7 +139,7 @@ async function createTask(req, res, next) {
         const teamSelected = await Team.findOne({ idTeam: taskNewData.assigned_team.idTeam })
   
         if (!teamSelected) {
-          logger.info('teamSelected not found.')
+          logger.error('teamSelected not found.')
           return res.status(400).end()
         }
         req.body.assigned_team = teamSelected._id
@@ -150,30 +150,30 @@ async function createTask(req, res, next) {
         const newUserAssigned = await User.findById(taskNewData.assigned_user)
   
         if (!newUserAssigned) {
-          logger.info('newUserAssigned not found.')
+          logger.error('newUserAssigned not found.')
           return res.status(400).end()
         }
         req.body.assigned_user = newUserAssigned._id
       }
 
       //valido el usuario autorizador
-      logger.info("authorized_by " + taskNewData.authorized_by)
+      //logger.info("authorized_by " + taskNewData.authorized_by)
       if (taskNewData.authorized_by) {
         const userAuthorized = await User.findById(taskNewData.authorized_by)
   
         if (!userAuthorized) {
-          logger.info('userAuthorized not found.')
+          logger.error('userAuthorized not found.')
           return res.status(400).end()
         }
         req.body.authorized_by = userAuthorized._id
       }else{
-        logger.info('userAuthorized not found.')
+        logger.error('userAuthorized not found.')
         return res.status(400).end()
       }
     
   
       await taskToUpdate.updateOne(req.body)
-      logger.info(taskToUpdate)
+      //logger.info(taskToUpdate)
       res.send(taskToUpdate)
 
     } catch (err) {
@@ -186,6 +186,10 @@ async function createTask(req, res, next) {
       return res.status(404).send('Parameter id not found')
     }
     try {
+       if (req.user.role.is_admin) {
+        logger.error('User is not admin.')
+        return res.status(403).end()
+      }
       //Valido tarea
       const taskDeleted = await Task.findOneAndDelete({_id : req.params.id});
       if (!taskDeleted) {
